qcacld-3.0: Fix dead loop when access per CPU TX/RX stats
NR_CPUS may be configured as 512 by CONFIG_NR_CPUS. Then uint8_t
is not a proper variable type to do for operation.

Change uint8_t to unsigned int to avoid dead loop.

CRs-Fixed: 3338268

+++ core/hdd/src/wlan_hdd_main.c
@@ -10304,7 +10304,7 @@ void hdd_send_mscs_action_frame(struct hdd_context *hdd_ctx,
 {
 	uint64_t mscs_vo_pkt_delta;
 	unsigned long tx_vo_pkts = 0;
-	uint8_t cpu;
+	unsigned int cpu;
 	struct hdd_tx_rx_stats *stats = &adapter->hdd_stats.tx_rx_stats;
 	uint32_t bus_bw_compute_interval;
